+++ packages/map-template/src/components/MapWrapper/MapWrapper.jsx
@@ -20,10 +20,9 @@ import solutionState from '../../atoms/solutionState';
 import notificationMessageState from '../../atoms/notificationMessageState';
 import useMapBoundsDeterminer from '../../hooks/useMapBoundsDeterminer';
 import hideNonMatchesState from '../../atoms/hideNonMatchesState';
-import miTransitionLevelState from '../../atoms/miTransitionLevelState';
-import showRoadNamesState from '../../atoms/showRoadNamesState';
 import PropTypes from 'prop-types';
 import ViewSelector from '../ViewSelector/ViewSelector';
+import LanguageSelector from '../LanguageSelector/LanguageSelector.jsx';
 import appConfigState from '../../atoms/appConfigState';
 import isNullOrUndefined from '../../helpers/isNullOrUndefined';
 
@@ -35,9 +34,10 @@ MapWrapper.propTypes = {
     onViewModeSwitchKnown: PropTypes.func.isRequired,
     resetCount: PropTypes.number.isRequired,
     mapOptions: PropTypes.object,
-    onMapOptionsChange: PropTypes.func,
     gmMapId: PropTypes.string,
-    isWayfindingOrDirections: PropTypes.bool
+    isWayfindingOrDirections: PropTypes.bool,
+    currentLanguage: PropTypes.string,
+    setLanguage: PropTypes.func
 };
 
 /**
@@ -58,12 +58,13 @@ let _tileStyle;
  * @param {function} onViewModeSwitchKnown - Function that is run when the view mode switch is known (if it is to be shown of not).
  * @param {number} resetCount - A counter that is incremented when the map should be reset.
  * @param {object} props.mapOptions - Options for instantiating and styling the map as well as UI elements.
- * @param {function} props.onMapOptionsChange - Function that is run when the map options are changed.
  * @param {string} props.gmMapId - Google Maps Map ID for custom styling.
  * @param {boolean} props.isWayfindingOrDirections - Whether wayfinding or directions is active or not.
+ * @param {string} props.currentLanguage - The currently selected language code.
+ * @param {function} props.setLanguage - Function to set the selected language.
  * @returns
  */
-function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, onMapOptionsChange, gmMapId, isWayfindingOrDirections }) {
+function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, gmMapId, isWayfindingOrDirections, currentLanguage, setLanguage }) {
     const apiKey = useRecoilValue(apiKeyState);
     const gmApiKey = useRecoilValue(gmApiKeyState);
     const mapboxAccessToken = useRecoilValue(mapboxAccessTokenState);
@@ -80,10 +81,9 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     const solution = useRecoilValue(solutionState);
     const [, setErrorMessage] = useRecoilState(notificationMessageState);
     const hideNonMatches = useRecoilValue(hideNonMatchesState);
-    const miTransitionLevel = useRecoilValue(miTransitionLevelState);
-    const showRoadNames = useRecoilValue(showRoadNamesState);
     const appConfig = useRecoilValue(appConfigState);
     const [isViewSelectorVisible, setIsViewSelectorVisible] = useState(false);
+    const [isLanguageSelectorVisible, setIsLanguageSelectorVisible] = useState(false);
 
     useLiveData(apiKey);
 
@@ -302,23 +302,7 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     }, [tileStyle]);
 
     /**
-     * React on changes in the miTransitionLevel prop.
-     */
-    useEffect(() => {
-        if (!isNaN(parseInt(miTransitionLevel))) {
-            onMapOptionsChange({ miTransitionLevel: miTransitionLevel })
-        }
-    }, [miTransitionLevel])
-
-    /**
-     * React on changes in the showRoadNames prop.
-     */
-    useEffect(() => {
-        onMapOptionsChange({ showRoadNames: showRoadNames })
-    }, [showRoadNames])
-
-    /**
-     * React on changes in appConfig and sets visibility of View Selector.
+     * React on changes in appConfig and sets visibility of View Selector and visibility of Language Selector.
      */
     useEffect(() => {
         if (appConfig) {
@@ -327,6 +311,14 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
             } else {
                 // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
                 setIsViewSelectorVisible(appConfig?.appSettings?.viewSelector.trim().toLowerCase() === 'true');
+            }
+
+
+            if (isNullOrUndefined(appConfig?.appSettings?.languageSelector)) {
+                setIsLanguageSelectorVisible(false);
+            } else {
+                // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
+                setIsLanguageSelectorVisible(appConfig?.appSettings?.languageSelector.trim().toLowerCase() === 'true');
             }
         }
     }, [appConfig])
@@ -342,7 +334,10 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
             gmMapId={gmMapId}
         />}
         {/* Pass isWayfindingOrDirections prop to ViewSelector to disable interactions while wayfinding or directions is active*/}
-        {apiKey && < ViewSelector isViewSelectorDisabled={isWayfindingOrDirections} isViewSelectorVisible={isViewSelectorVisible} />}
+        {apiKey && <>
+            <ViewSelector isViewSelectorVisible={isViewSelectorVisible} isViewSelectorDisabled={isWayfindingOrDirections} />
+            <LanguageSelector currentLanguage={currentLanguage} setLanguage={setLanguage} isVisible={isLanguageSelectorVisible} />
+        </>}
     </>)
 }