+++ packages/map-template/src/i18n/fr.js
@@ -81,6 +81,8 @@ const fr = {
     // View Selector
     'Go to view': 'Afficher le plan',
     'Close view selector': 'Fermer le sélecteur de vue',
+    // Language Selector
+    'Select language': 'Sélectionner langue',
 };
 
 export default fr;